# openapi/minimovie-api.yaml
openapi: 3.0.3
info:
  title: MiniMovie API
  description: API for retrieving intuitive movie, TV series, and person information
  version: 1.0.0
servers:
  - url: https://api.minimovie.info
    description: Production server
security:
  - bearerAuth: []
paths:
...
  /users/me/watchlist:
    get:
      summary: List watchlist
      description: Returns the user's watchlist items, optionally filtered by status or media type
      operationId: listWatchlist
      security:
        - sessionCookie: []
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum:
              - want_to_watch
              - watched
        - name: media_type
          in: query
          required: false
          schema:
            type: string
            enum:
              - movie
              - series
      responses:
        "200":
          description: Watchlist items
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WatchlistListResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      summary: Add to watchlist
      description: Adds a movie or series to the user's watchlist, resolving metadata from TMDB
      operationId: addToWatchlist
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddWatchlistRequest"
      responses:
        "201":
          description: Item added to watchlist
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WatchlistItem"
        "400":
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Media not found on TMDB
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Item already in watchlist
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "503":
          description: Metadata resolution failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /users/me/watchlist/check:
    get:
      summary: Check if item is in watchlist
      description: Checks whether a specific media item exists in the user's watchlist
      operationId: checkWatchlist
      security:
        - sessionCookie: []
      parameters:
        - name: media_type
          in: query
          required: true
          schema:
            type: string
            enum:
              - movie
              - series
        - name: media_id
          in: query
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Check result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WatchlistCheckResponse"
        "400":
          description: Invalid media_id
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /users/me/watchlist/{id}:
    patch:
      summary: Update watchlist item status
      description: Updates the status of a watchlist item
      operationId: updateWatchlistStatus
      security:
        - sessionCookie: []
      parameters:
        - name: id
          in: path
          required: true
          description: Watchlist item ID
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateWatchlistStatusRequest"
      responses:
        "200":
          description: Updated watchlist item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WatchlistItem"
        "400":
          description: Invalid status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Item not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      summary: Remove from watchlist
      description: Removes an item from the user's watchlist
      operationId: removeFromWatchlist
      security:
        - sessionCookie: []
      parameters:
        - name: id
          in: path
          required: true
          description: Watchlist item ID
          schema:
            type: string
      responses:
        "204":
          description: Item removed
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Item not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Guest API key used for public catalog routes and token exchange
    sessionCookie:
      type: apiKey
      in: cookie
      name: mm_session
      description: Session cookie set after OAuth login (production name is __Host-mm_session)
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: HTTP status text
        message:
          type: string
          description: Detailed error message
...
    WatchlistItem:
      type: object
      properties:
        id:
          type: string
        mediaType:
          type: string
          enum:
            - movie
            - series
        mediaId:
          type: integer
        mediaTitle:
          type: string
        posterPath:
          type: string
          nullable: true
        status:
          type: string
          enum:
            - want_to_watch
            - watched
        startedAt:
          type: string
          format: date-time
          nullable: true
        finishedAt:
          type: string
          format: date-time
          nullable: true
        lastWatchedAt:
          type: string
          format: date-time
          nullable: true
        watchCount:
          type: integer
        genres:
          type: array
          items:
            type: string
        runtimeMinutes:
          type: integer
          nullable: true
        voteAverage:
          type: number
          format: float
          nullable: true
        releaseYear:
          type: integer
          nullable: true
        addedAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    WatchlistListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/WatchlistItem"
        total:
          type: integer
    WatchlistCheckResponse:
      type: object
      properties:
        exists:
          type: boolean
        item:
          type: object
          description: Only present when exists is true
          properties:
            id:
              type: string
            status:
              type: string
    AddWatchlistRequest:
      type: object
      required:
        - mediaType
        - mediaId
        - status
      properties:
        mediaType:
          type: string
          enum:
            - movie
            - series
        mediaId:
          type: integer
          minimum: 1
        status:
          type: string
          enum:
            - want_to_watch
            - watched
    UpdateWatchlistStatusRequest:
      type: object
      required:
        - status
      properties:
        status:
          type: string
          enum:
            - want_to_watch
            - watched
...